+++ scratch/bmm_profile.py
@@ -76,11 +76,13 @@ def main():
     device = torch.device("cuda")
 
     # top_scores: (tokens, top_k), routed_output: (tokens, top_k, dim)
+    # For the MoE BMM setup the scores are in float32 and the routed_output is usually in
+    # low-precision.
     top_scores = torch.randn(
         tokens, top_k, device=device, dtype=torch.float32, requires_grad=True
     )
     routed_output = torch.randn(
-        tokens, top_k, dim, device=device, dtype=torch.float32, requires_grad=True
+        tokens, top_k, dim, device=device, dtype=torch.bfloat16, requires_grad=True
     )
 
     # Warmup
